Make listing prices integers

+++ core/models.py
@@ -624,9 +624,9 @@ class Listing(models.Model):
     )
     description = models.TextField()
 
-    original_price = models.DecimalField(max_digits=10, decimal_places=2, default=0)
-    listing_price = models.DecimalField(max_digits=10, decimal_places=2, default=0)
-    lowest_price = models.DecimalField(max_digits=10, decimal_places=2, default=250)
+    original_price = models.PositiveSmallIntegerField(default=0)
+    listing_price = models.PositiveSmallIntegerField(default=0)
+    lowest_price = models.PositiveSmallIntegerField(default=250)
 
     campaign = models.ForeignKey(
         Campaign,
